Fix qp signalling type
Per the IB spec for create qp(spec 1.2.1 section 11.2.4.11) The Signaling
Type must be specified for the Send Queue on this QP.  The valid
types are:

* Non-selectable: All Work Requests submitted to the Send Queue always
  generate a completion entry.

* Selectable: Consumer must specify on each Work Request submitted to
  the Send Queue whether to generate a completion entry for successful
  completions.

The default for IB is non-selectable (.sq_sig_type set to 0).  In the
qp_init_attr structure the field .sq_sig_type needs to be set to
IB_SIGNAL_REQ_WR in order to set it as selecetable (ie controlled from
the WR).

Signed-off-by: Jon Mason <jon@opengridcomputing.com>
Signed-off-by: Steve Wise <swise@opengridcomputing.com>